Implement transform stage in pure Elixir
This is actually the initial implementation before trying
out the Rust NIF transform. This is also faster since it
does not bother chunking the slices into valid and invalid
string parts only to pass again to the Rust NIF.

This also removes the partition stage as it is also not
useful for this workload as the reconciliation of partitioned
case is the same as the non-partitioned case. So the cost
of hashing to route to the same partitions add no real value.